+++ services/backend/src/database/migrations/2021.06.17-1503-addExpireAtAttributeToTestRedeemsTable.js
@@ -1,0 +1,11 @@
+module.exports = {
+  up: async (queryInterface, Sequelize) => {
+    await queryInterface.addColumn('TestRedeems', 'expireAt', {
+      type: Sequelize.DATE,
+    });
+  },
+
+  down: async queryInterface => {
+    await queryInterface.removeColumn('TestRedeems', 'expireAt');
+  },
+};