+++ services/backend/src/database/models/testRedeems.js
@@ -9,6 +9,9 @@ module.exports = (Sequelize, DataTypes) => {
       type: DataTypes.STRING(24),
       allowNull: false,
     },
+    expireAt: {
+      type: DataTypes.DATE,
+    },
   });
 
   TestRedeem.associate = () => {};